Updating ResourceManager to allow multiple languages to be specified (for vs-mda-remote) Adding tests for ResourceManger (which pass!)
Added more error checking to typescript compilation (When the error is not to do with a file, it would crash in the error reporting code).
Adding in more utility code

Added in cordova config parser to utils, along with the start of some file helpers.
Fixed up compilation of resources.ts. Protip: ///references have to be before any code, such as "use strict"

Added BuildInfo class to util since that will be common across client and server (as well as server components)