Trying out alternate approach to taco-utils
We can split up the TacoUtility module across multiple files, but we need to manually ensure that the mapping between what the generated code does and what the type system says is correct.

The approach that I'm trying here should work fine UNLESS we need types from other node modules to be part of the type system; e.g. if we had a function that took an express.Request or returned a Q.Promise then it will need a different approach.