Fixing remote plugins with parameters
With the change to always plugin-add plugins, we missed the case when there were parameters to those plugins. For correctly-behaving cordova, these plugins are saved to fetch.json and we can pull them from there. For older versions, VS saves the parameters to config.xml so we can read from there as well.