Fixing remote plugins with parameters
With the change to always plugin-add plugins, we missed the case when there were parameters to those plugins. For correctly-behaving cordova, these plugins are saved to fetch.json and we can pull them from there. For incorrect versions of cordova (such as 4.3.1) we cannot support this scenario since adding a platform after adding a plugin with parameters will fail.