Reducing likelyhood of taco-cli certificate error
There seems to be some kind of race condition in node 10.x and 12.x. When the stars align, even if you've made the same https request 10 times before, suddenly this one can blow up and give you an uncaught (and un-catchable) exception.

There seems to be a fix in io.js which might make it back into node one day, but for now the best way to deal with things looks like playing with timing. It seems that adding a slight delay reduces the likelyhood of this issue coming up.