Starting work to make auto-updating packages more reliable
Right now if you lose internet, or your temporary NPM cache is not accessable for dev scenarios, we will fail out as soon as we try and update even though we could continue working with a perfectly good (but potentially outdated) package. With this first change we'll no longer automatically fail out when the local cache is not present, but instead we'll fail after the timeout expires and we try to get a new version.

Next is to modify the "get a new version" code to distinguish between "failed to update" and "failed to install", and making "failred to update" not trash the previous goood version.