Adding in checks for valid platforms where possible
Now when we build or run, we try to verify platforms against known cordova platforms, and ignore platforms that are not known. This is to try and avoid issues with remotebuild potentially trying to build newer platforms in older cordovas.

Due to limitations on the "cordova platforms" CLI command, we can't get this information out of the CLI and so when not in a kit project, we assume everything is a valid project and don't get these benefits.

Still need to make the changes to "taco setup remote" to avoid adding unknown platforms when possible.