Improving CordovaWrapper
CordovaWrapper now has a simpler pattern across functions: they invoke cordovaApiOrProcess with two functions: one to invoke cordova APIs on an ICordova which is passed in, and one to generate a command line for the cordova CLI interface.

As part of this, I made slight changes to kits, so they now consult taco.json (after writing it) to find the version of cordova to use, rather than passing that in directly. This allows for a unified interface.