Improving resiliency around iOS device launch
There was the potential for socket communication between us and the iOS device to have problems that we weren't catching, which could take down the whole server. We now listen to the error event of the communication, which should prevent that.
I also tried to fix up some error reporting problems that I encountered while trying to cause network communication issues.

Also fixing up how deleted files are handled, to make sure that platform-specific path separators are properly handled and we also don't allow files outside of the project directory to be deleted.