Increasing HTTP connection timeouts
When uploading larger projects, some connections would be terminated unexpectedly.
I believe that it is related to https://blog.cloudflare.com/the-curious-case-of-slow-downloads/
and so I have increased the timout to work around that. If it is the linked issue, then a 5 minute timeout should be sufficient to allow arbitrary uploads.